Dockerize Laravel
I basically followed a tutorial on digital ocean to dockerize the
application. The one change I have made from the tutorial is that the
docker-compose.yml file references environment variables instead of
having the mysql root password and database name in the file itself.

The purpose of this commit is to facilitate development work. A thing I
find missing is the manual step of shelling into the db container and
granting the DB_USER access to the database. Will look into possibly
automating that in the future as well.

Here is the link to the digital ocean article
https://www.digitalocean.com/community/tutorials/how-to-set-up-laravel-nginx-and-mysql-with-docker-compose